# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Views Rocket App API",
    "description": "Node.js and Express.js Application API",
    "license": {
      "name": "Views rocket App"
    }
  },
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "jwt": {
      "description": "JWT token format : Bearer [JWT token] ",
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "security": [
    {
      "jwt": []
    }
  ],
  "paths": {
...
    "/campaign": {
      "post": {
        "tags": ["Campaign"],
        "description": "Api for create Campaign in app",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create Campaign",
            "schema": {
              "$ref": "#/definitions/Campaign"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User log in response",
            "schema": {
              "$ref": "#/definitions/Campaign"
            }
          }
        }
      },
      "get": {
        "tags": ["Campaign"],
        "summary": "Get all Campaign",
        "responses": {
          "200": {
            "description": "Tree structured JSON formatted response for Campaign"
          }
        }
      }
    },
...
    "/campaign/stopCampaign": {
      "post": {
        "tags": ["Campaign"],
        "description": "Api for stop Campaign in app",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to stop Campaign",
            "schema": {
              "$ref": "#/definitions/stopCampaign"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "stop Campaign response",
            "schema": {
              "$ref": "#/definitions/stopCampaign"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Campaign": {
      "required": ["videoURL", "desiredViewcount","desiredViewduration","cost"],
      "properties": {
        "videoURL": {
          "type": "string"
        },
        "desiredViewcount": {
          "type": "number"
        },
        "desiredViewduration": {
          "type": "number",
          "description": "How many seconds you want"
        },
        "cost": {
          "type": "number",
          "description": "Must be Grater then equal to desiredView * Viewduration"
        }
      }
    },
    "stopCampaign": {
      "required": ["campaignId"],
      "properties": {
        "campaignId": {
          "type": "string"
        }
      }
    },
...
  }
}